Move analyse command logic to pkg, and export it, allowing it to be used as a module in other projects
